Labaratory work #3

+++ laboratory work/src/main/java/org/dut/lab3/Product.java
@@ -2,5 +2,6 @@ package org.dut.lab3;
 
 import java.math.BigDecimal;
 
-public record Product(int id, String name, BigDecimal price) {
+public record Product(int id, String name, double price) {
+
 }
